server-requirements.txt
python-dotenv==0.15.0
discord.py==1.6.0